Minor README improvement.
I added Git as an optional requirement, which would be needed if the user intended on cloning the repository through their Terminal. It is not a requirement if the user decides to download the repository through their browser (That has also been mentioned).

Not sure if this is actually needed but it would make things slightly clearer for people who aren't as experienced.